Refactor : 그룹 수정 페이지일때, 상세 데이터 가져오는 로직을 커스텀 훅으로 분리
