+++ src/pages/group/GroupEdit/index.tsx
@@ -1,13 +1,12 @@
 import { useEffect, useState } from 'react';
+import { useLocation } from 'react-router-dom';
 
 import GroupNavSidebar from '@/components/group/recruit/GroupNavSidebar';
 import GroupRegistInfo from '@/components/group/recruit/GroupRegistInfo';
 import GroupRegistSchedule from '@/components/group/recruit/GroupRegistSchedule';
 import KakaoMap from '@/components/group/recruit/KakaoMap';
-import { useLocation } from 'react-router-dom';
-import { useRecoilState } from 'recoil';
-import { selectedPlaceState, tempFormState } from '@/recoil/GroupRegistState';
-import { fetchGroupDetail } from '@/apis/groupDetail';
+
+import { useEditGroup } from '@/hooks/useRegist';
 
 function GroupEdit() {
   const [currentStep, setCurrentStep] = useState(1); // Step1 : 기본 정보, Step2 : 일정 선택
@@ -17,52 +16,10 @@ function GroupEdit() {
   };
 
   const { pathname } = useLocation();
-  const [tempFormData, setTempFormData] = useRecoilState(tempFormState);
-  const [selectedPlace, setSelectedPlace] = useRecoilState(selectedPlaceState);
+  const { handleEditData } = useEditGroup();
 
   useEffect(() => {
-    const fetchData = async () => {
-      if (pathname.startsWith('/group/edit')) {
-        const match = pathname.match(/\/(\d+)$/);
-
-        if (match) {
-          const groupId = match[1];
-          const response = await fetchGroupDetail(Number(groupId));
-          console.log(response);
-
-          const preProcessData = response.places.map((item: any) => {
-            return {
-              placeCode: item.placeCode,
-              name: item.name,
-              address: item.address,
-              category: item.category,
-              orders: item.orders,
-              x: Number(item.x),
-              y: Number(item.y),
-            };
-          });
-
-          console.log(preProcessData);
-          setSelectedPlace({ selectedPlace: preProcessData });
-
-          console.log(response);
-          const tempFormData = {
-            groupTitle: response.tripGroupDetail.name,
-            groupDescription: response.tripGroupDetail.content,
-            groupMemberCount: response.tripGroupDetail.maxUserNumber,
-            departureDate: new Date(response.tripGroupDetail.tripDate),
-            recruitmentEndDate: new Date(response.tripGroupDetail.dueDate),
-            groupThumbnail: response.tripGroupDetail.thumbnailUrl,
-          };
-
-          setTempFormData(tempFormData);
-        } else {
-          console.log('No match found');
-        }
-      }
-    };
-
-    fetchData();
+    handleEditData(pathname);
   }, [pathname]);
 
   return (
